cleanup and AC #1,2
